+++ config/initializers/01_settings.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+BootConfig.load_from_env!
+ApplicationConfig.load_from_env!